lsp configuraion with lazy.nvim
